requirements.txt
scrapy==2.13.0
scikit-learn==1.6.0
flask==3.0.3
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
jsonschema==4.22.0
faiss-cpu==1.7.4
gensim==4.3.2
pytest==8.2.1
pytest-cov==4.1.0
gunicorn==21.2.0
jieba==0.42.1
html2text==2020.1.16
tqdm==4.66.2
pyspellchecker==0.7.2